Put individual ModelTests into their own files.
