+++ WalletWasabi.Tests/UnitTests/AllFeeEstimateTests.cs
@@ -1,0 +1,32 @@
+using NBitcoin.RPC;
+using Newtonsoft.Json;
+using System;
+using System.Collections.Generic;
+using System.Text;
+using WalletWasabi.Models;
+using Xunit;
+
+namespace WalletWasabi.Tests.UnitTests
+{
+	public class AllFeeEstimateTests
+	{
+		[Fact]
+		public void AllFeeEstimateSerialization()
+		{
+			var estimations = new Dictionary<int, int>
+			{
+				{ 2, 102 },
+				{ 3, 20 },
+				{ 19, 1 }
+			};
+			var allFee = new AllFeeEstimate(EstimateSmartFeeMode.Conservative, estimations);
+			var serialized = JsonConvert.SerializeObject(allFee);
+			var deserialized = JsonConvert.DeserializeObject<AllFeeEstimate>(serialized);
+
+			Assert.Equal(estimations[2], deserialized.Estimations[2]);
+			Assert.Equal(estimations[3], deserialized.Estimations[3]);
+			Assert.Equal(estimations[19], deserialized.Estimations[19]);
+			Assert.Equal(EstimateSmartFeeMode.Conservative, deserialized.Type);
+		}
+	}
+}
